+++ opentelemetry-sdk/src/opentelemetry/sdk/trace/__init__.py
@@ -47,7 +47,7 @@ from opentelemetry.trace.propagation import SPAN_KEY
 from opentelemetry.trace.status import (
     EXCEPTION_STATUS_FIELD,
     Status,
-    StatusCanonicalCode,
+    StatusCode,
 )
 from opentelemetry.util import time_ns, types
 
@@ -373,11 +373,15 @@ def _create_immutable_attributes(attributes):
 
 def _check_span_ended(func):
     def wrapper(self, *args, **kwargs):
+        already_ended = False
         with self._lock:  # pylint: disable=protected-access
-            if self.end_time is not None:
-                logger.warning("Calling %s on an ended span.", func.__name__)
-                return
-            func(self, *args, **kwargs)
+            if self.end_time is None:
+                func(self, *args, **kwargs)
+            else:
+                already_ended = True
+
+        if already_ended:
+            logger.warning("Tried calling %s on an ended span.", func.__name__)
 
     return wrapper
 
@@ -408,6 +412,7 @@ class Span(trace_api.Span):
             raise TypeError("Span must be instantiated via a tracer.")
         return super().__new__(cls)
 
+    # pylint: disable=too-many-locals
     def __init__(
         self,
         name: str,
@@ -422,6 +427,7 @@ class Span(trace_api.Span):
         kind: trace_api.SpanKind = trace_api.SpanKind.INTERNAL,
         span_processor: SpanProcessor = SpanProcessor(),
         instrumentation_info: InstrumentationInfo = None,
+        record_exception: bool = True,
         set_status_on_exception: bool = True,
     ) -> None:
 
@@ -432,10 +438,11 @@ class Span(trace_api.Span):
         self.trace_config = trace_config
         self.resource = resource
         self.kind = kind
+        self._record_exception = record_exception
         self._set_status_on_exception = set_status_on_exception
 
         self.span_processor = span_processor
-        self.status = None
+        self.status = Status(StatusCode.UNSET)
         self._lock = threading.Lock()
 
         _filter_attribute_values(attributes)
@@ -546,7 +553,7 @@ class Span(trace_api.Span):
 
         if self.status is not None:
             status = OrderedDict()
-            status["canonical_code"] = str(self.status.canonical_code.name)
+            status["status_code"] = str(self.status.status_code.name)
             if self.status.description:
                 status["description"] = self.status.description
 
@@ -637,9 +644,6 @@ class Span(trace_api.Span):
                 logger.warning("Calling end() on an ended span.")
                 return
 
-            if self.status is None:
-                self.status = Status(canonical_code=StatusCanonicalCode.OK)
-
             self._end_time = end_time if end_time is not None else time_ns()
 
         self.span_processor.on_end(self)
@@ -649,7 +653,7 @@ class Span(trace_api.Span):
         self.name = name
 
     def is_recording(self) -> bool:
-        return True
+        return self._end_time is None
 
     @_check_span_ended
     def set_status(self, status: trace_api.Status) -> None:
@@ -662,22 +666,35 @@ class Span(trace_api.Span):
         exc_tb: Optional[TracebackType],
     ) -> None:
         """Ends context manager and calls `end` on the `Span`."""
-
-        if (
-            self.status is None
-            and self._set_status_on_exception
-            and exc_val is not None
-        ):
-            self.set_status(
-                Status(
-                    canonical_code=StatusCanonicalCode.UNKNOWN,
-                    description="{}: {}".format(exc_type.__name__, exc_val),
+        if exc_val is not None:
+            # Record the exception as an event
+            # pylint:disable=protected-access
+            if self._record_exception:
+                self.record_exception(exception=exc_val, escaped=True)
+            # Records status if span is used as context manager
+            # i.e. with tracer.start_span() as span:
+            if (
+                self.status.status_code is StatusCode.UNSET
+                and self._set_status_on_exception
+            ):
+                self.set_status(
+                    Status(
+                        status_code=StatusCode.ERROR,
+                        description="{}: {}".format(
+                            exc_type.__name__, exc_val
+                        ),
+                    )
                 )
-            )
 
         super().__exit__(exc_type, exc_val, exc_tb)
 
-    def record_exception(self, exception: Exception) -> None:
+    def record_exception(
+        self,
+        exception: Exception,
+        attributes: types.Attributes = None,
+        timestamp: Optional[int] = None,
+        escaped: bool = False,
+    ) -> None:
         """Records an exception as a span event."""
         try:
             stacktrace = traceback.format_exc()
@@ -686,14 +703,16 @@ class Span(trace_api.Span):
             # an AttributeError if the __context__ on
             # an exception is None
             stacktrace = "Exception occurred on stacktrace formatting"
-
+        _attributes = {
+            "exception.type": exception.__class__.__name__,
+            "exception.message": str(exception),
+            "exception.stacktrace": stacktrace,
+            "exception.escaped": str(escaped),
+        }
+        if attributes:
+            _attributes.update(attributes)
         self.add_event(
-            name="exception",
-            attributes={
-                "exception.type": exception.__class__.__name__,
-                "exception.message": str(exception),
-                "exception.stacktrace": stacktrace,
-            },
+            name="exception", attributes=_attributes, timestamp=timestamp
         )
 
 
@@ -706,19 +725,22 @@ class _Span(Span):
 
 class Tracer(trace_api.Tracer):
     """See `opentelemetry.trace.Tracer`.
-
-    Args:
-        name: The name of the tracer.
-        shutdown_on_exit: Register an atexit hook to shut down the tracer when
-            the application exits.
     """
 
     def __init__(
         self,
-        source: "TracerProvider",
+        sampler: sampling.Sampler,
+        resource: Resource,
+        span_processor: Union[
+            SynchronousMultiSpanProcessor, ConcurrentMultiSpanProcessor
+        ],
+        ids_generator: trace_api.IdsGenerator,
         instrumentation_info: InstrumentationInfo,
     ) -> None:
-        self.source = source
+        self.sampler = sampler
+        self.resource = resource
+        self.span_processor = span_processor
+        self.ids_generator = ids_generator
         self.instrumentation_info = instrumentation_info
 
     def start_as_current_span(
@@ -728,12 +750,21 @@ class Tracer(trace_api.Tracer):
         kind: trace_api.SpanKind = trace_api.SpanKind.INTERNAL,
         attributes: types.Attributes = None,
         links: Sequence[trace_api.Link] = (),
+        start_time: Optional[int] = None,
         record_exception: bool = True,
+        set_status_on_exception: bool = True,
     ) -> Iterator[trace_api.Span]:
-        span = self.start_span(name, context, kind, attributes, links)
-        return self.use_span(
-            span, end_on_exit=True, record_exception=record_exception
+        span = self.start_span(
+            name=name,
+            context=context,
+            kind=kind,
+            attributes=attributes,
+            links=links,
+            start_time=start_time,
+            record_exception=record_exception,
+            set_status_on_exception=set_status_on_exception,
         )
+        return self.use_span(span, end_on_exit=True)
 
     def start_span(  # pylint: disable=too-many-locals
         self,
@@ -743,6 +774,7 @@ class Tracer(trace_api.Tracer):
         attributes: types.Attributes = None,
         links: Sequence[trace_api.Link] = (),
         start_time: Optional[int] = None,
+        record_exception: bool = True,
         set_status_on_exception: bool = True,
     ) -> trace_api.Span:
 
@@ -760,7 +792,7 @@ class Tracer(trace_api.Tracer):
         # is_valid determines root span
         if parent_span_context is None or not parent_span_context.is_valid:
             parent_span_context = None
-            trace_id = self.source.ids_generator.generate_trace_id()
+            trace_id = self.ids_generator.generate_trace_id()
             trace_flags = None
             trace_state = None
         else:
@@ -773,8 +805,8 @@ class Tracer(trace_api.Tracer):
         # exported.
         # The sampler may also add attributes to the newly-created span, e.g.
         # to include information about the sampling result.
-        sampling_result = self.source.sampler.should_sample(
-            context, trace_id, name, attributes, links,
+        sampling_result = self.sampler.should_sample(
+            context, trace_id, name, attributes, links, trace_state
         )
 
         trace_flags = (
@@ -784,10 +816,10 @@ class Tracer(trace_api.Tracer):
         )
         span_context = trace_api.SpanContext(
             trace_id,
-            self.source.ids_generator.generate_span_id(),
+            self.ids_generator.generate_span_id(),
             is_remote=False,
             trace_flags=trace_flags,
-            trace_state=trace_state,
+            trace_state=sampling_result.trace_state,
         )
 
         # Only record if is_recording() is true
@@ -797,13 +829,14 @@ class Tracer(trace_api.Tracer):
                 name=name,
                 context=span_context,
                 parent=parent_span_context,
-                sampler=self.source.sampler,
-                resource=self.source.resource,
+                sampler=self.sampler,
+                resource=self.resource,
                 attributes=sampling_result.attributes.copy(),
-                span_processor=self.source._active_span_processor,
+                span_processor=self.span_processor,
                 kind=kind,
                 links=links,
                 instrumentation_info=self.instrumentation_info,
+                record_exception=record_exception,
                 set_status_on_exception=set_status_on_exception,
             )
             span.start(start_time=start_time, parent_context=context)
@@ -813,10 +846,7 @@ class Tracer(trace_api.Tracer):
 
     @contextmanager
     def use_span(
-        self,
-        span: trace_api.Span,
-        end_on_exit: bool = False,
-        record_exception: bool = True,
+        self, span: trace_api.Span, end_on_exit: bool = False,
     ) -> Iterator[trace_api.Span]:
         try:
             token = context_api.attach(context_api.set_value(SPAN_KEY, span))
@@ -825,22 +855,24 @@ class Tracer(trace_api.Tracer):
             finally:
                 context_api.detach(token)
 
-        except Exception as error:  # pylint: disable=broad-except
-            # pylint:disable=protected-access
+        except Exception as exc:  # pylint: disable=broad-except
+            # Record the exception as an event
             if isinstance(span, Span):
-                if record_exception:
-                    span.record_exception(error)
+                # pylint:disable=protected-access
+                if span._record_exception:
+                    span.record_exception(exc)
 
-                if span.status is None and span._set_status_on_exception:
+                # Records status if use_span is used
+                # i.e. with tracer.start_as_current_span() as span:
+                if (
+                    span.status.status_code is StatusCode.UNSET
+                    and span._set_status_on_exception
+                ):
                     span.set_status(
                         Status(
-                            canonical_code=getattr(
-                                error,
-                                EXCEPTION_STATUS_FIELD,
-                                StatusCanonicalCode.UNKNOWN,
-                            ),
+                            status_code=StatusCode.ERROR,
                             description="{}: {}".format(
-                                type(error).__name__, error
+                                type(exc).__name__, exc
                             ),
                         )
                     )
@@ -884,7 +916,10 @@ class TracerProvider(trace_api.TracerProvider):
             instrumenting_module_name = "ERROR:MISSING MODULE NAME"
             logger.error("get_tracer called with missing module name.")
         return Tracer(
-            self,
+            self.sampler,
+            self.resource,
+            self._active_span_processor,
+            self.ids_generator,
             InstrumentationInfo(
                 instrumenting_module_name, instrumenting_library_version
             ),