+++ opentelemetry-proto/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Update protobuf versions
+  ([#1356](https://github.com/open-telemetry/opentelemetry-python/pull/1356))
+
 ## Version 0.13b0
 
 Released 2020-09-17